backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
websockets==12.0
yt-dlp==2023.11.16
google-api-python-client==2.108.0
spotipy==2.23.0
python-jose==3.3.0
cloudflared==2024.2.0 